Split MethodSimulator out of MethodMetadata; add tabarena_public preset
Stage 4 of the MethodMetadata refinement: separate simulation behavior from the
metadata record, and replace the repeated public-artifact kwargs with a factory.

Record/behavior split: the five PaperRunTabArena-based methods (get_config_default,
generate_results, generate_hpo_result, generate_hpo_trajectories, generate_best)
move from MethodMetadata into a new MethodSimulator (models/_method_simulator.py)
that wraps a MethodMetadata. Callers (end_to_end_single, abstract_arena_context,
and the calibration/trajectory scripts) now go through MethodSimulator(mm).x(...).
generate_repo / generate_repo_holdout / load_hpo_trajectories stay on the record.

Preset factory: MethodMetadata.tabarena_public(...) supplies the public-TabArena
defaults (s3_bucket="tabarena", s3_prefix="cache", upload_as_public=True,
has_raw/has_processed/has_results=True), overridable per call. baselines/info.py and
the seven dated _tabarena_method_metadata_*.py artifact files now build through it,
shrinking/removing their _common_*_kwargs storage boilerplate.

Behavior-preserving: to_info_dict() for all 121 production MethodMetadata objects is
byte-for-byte identical; the 5 methods are gone from MethodMetadata and present on
MethodSimulator with no stray direct callers.
