+++ packages/tabarena/src/tabarena/nips2025_utils/artifacts/_tabarena_method_metadata_2025_06_12.py
@@ -43,16 +43,11 @@ common_kwargs = dict(
     artifact_name="tabarena-2025-06-12",
     date="2025-06-12",
     method_type="config",
-    has_raw=True,
-    has_processed=True,
-    has_results=True,
     # FIXME: technically GBDTs and baselines are not verified
     verified=True,
 )
 
 s3_cache_kwargs = dict(
-    s3_bucket="tabarena",
-    s3_prefix="cache",
     cache_type="s3",
 )
 
@@ -312,7 +307,7 @@ for method in methods:
 
     method_kwargs = copy.deepcopy(method_kwargs)
 
-    method_metadata = MethodMetadata(
+    method_metadata = MethodMetadata.tabarena_public(
         method=method,
         config_default=config_default,
         display_name=display_name,
