+++ packages/tabarena/src/tabarena/nips2025_utils/artifacts/_tabarena_method_metadata_2025_10_20.py
@@ -4,17 +4,13 @@ from tabarena.models._method_metadata import MethodMetadata
 
 _common_kwargs = dict(
     artifact_name="tabarena-2025-10-20",
-    s3_bucket="tabarena",
-    s3_prefix="cache",
     cache_type="s3",
-    upload_as_public=True,
-    has_results=True,
     name_suffix=None,
     # FIXME: technically LR and kNN are not verified
     verified=True,
 )
 
-portfolio_metadata_paper_cr = MethodMetadata(
+portfolio_metadata_paper_cr = MethodMetadata.tabarena_public(
     method="Portfolio-N200-4h",
     method_type="portfolio",
     date="2025-10-20",
