+++ packages/tabarena/src/tabarena/nips2025_utils/artifacts/_tabarena_method_metadata_2025_12_18.py
@@ -4,13 +4,7 @@ from tabarena.models._method_metadata import MethodMetadata
 
 _common_kwargs = dict(
     artifact_name="tabarena-2025-12-18",
-    s3_bucket="tabarena",
-    s3_prefix="cache",
     cache_type="s3",
-    upload_as_public=True,
-    has_raw=True,
-    has_processed=True,
-    has_results=True,
     method_type="baseline",
     name_suffix=None,
     date="2025-12-18",
@@ -28,7 +22,7 @@ _cpu_kwargs = dict(
     **_common_kwargs,
 )
 
-ag_150_eq_4h8c_metadata = MethodMetadata(
+ag_150_eq_4h8c_metadata = MethodMetadata.tabarena_public(
     method="AutoGluon_v150_eq_4h8c",
     name="AutoGluon 1.5 (extreme, 4h)",
     **_gpu_kwargs,
