+++ packages/tabarena/src/tabarena/nips2025_utils/artifacts/_tabarena_method_metadata_misc.py
@@ -4,7 +4,7 @@ from tabarena.models._method_metadata import MethodMetadata
 
 # LightGBM w/ custom preprocessing pipeline (only first 3 repeats)
 # s3 cache = "cache_aio"
-gbm_aio_0808_metadata = MethodMetadata(
+gbm_aio_0808_metadata = MethodMetadata.tabarena_public(
     method="LightGBM_aio_0808",
     artifact_name="LightGBM_aio_0808",
     method_type="config",
@@ -14,13 +14,8 @@ gbm_aio_0808_metadata = MethodMetadata(
     model_key="GBM_aio_0808",
     config_default="LightGBM_aio_0808_c1_BAG_L1",
     name_suffix=None,
-    has_raw=True,
-    has_processed=True,
-    has_results=True,
-    upload_as_public=True,
     can_hpo=True,
     is_bag=True,
-    s3_bucket="tabarena",
     s3_prefix="cache_aio",
     cache_type="s3",
     verified=False,
@@ -28,7 +23,7 @@ gbm_aio_0808_metadata = MethodMetadata(
 
 # LightGBM w/ custom preprocessing pipeline (only first 3 repeats)
 # s3 cache = "cache_aio"
-prep_gbm_v6_metadata = MethodMetadata(
+prep_gbm_v6_metadata = MethodMetadata.tabarena_public(
     method="prep_LightGBM_v6",
     artifact_name="prep_LightGBM_v6",
     method_type="config",
@@ -38,13 +33,8 @@ prep_gbm_v6_metadata = MethodMetadata(
     model_key="prep_GBM_v6",
     config_default="prep_LightGBM_v6_c1_BAG_L1",
     name_suffix=None,
-    has_raw=True,
-    has_processed=True,
-    has_results=True,
-    upload_as_public=True,
     can_hpo=True,
     is_bag=True,
-    s3_bucket="tabarena",
     s3_prefix="cache_aio",
     cache_type="s3",
     verified=True,
